fix: credhub_setup: move packages around more
Actually do what the review asked, and move things into cmd/ and pkg/
