+++ deploy/containers/credhub_setup/cmd/BUILD.bazel
@@ -1,0 +1,23 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["main.go"],
+    importpath = "credhub_setup/cmd",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//pkg/cc:go_default_library",
+        "//pkg/credhub:go_default_library",
+        "//pkg/uaa:go_default_library",
+    ],
+)
+
+go_binary(
+    name = "cmd",
+    embed = [":go_default_library"],
+    goarch = "amd64",
+    goos = "linux",
+    pure = "on",
+    static = "on",
+    visibility = ["//visibility:public"],
+)
