+++ deploy/containers/credhub_setup/pkg/cc/link_test.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"credhub_setup/cc"
-	cchelpers "credhub_setup/cc/testhelpers"
-	quarkshelpers "credhub_setup/quarks/testhelpers"
+	"credhub_setup/pkg/cc"
+	cchelpers "credhub_setup/pkg/cc/testhelpers"
+	quarkshelpers "credhub_setup/pkg/quarks/testhelpers"
 )
 
 func TestNewHTTPClient(t *testing.T) {
