+++ deploy/containers/credhub_setup/pkg/cc/testhelpers/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["mockserver.go"],
-    importpath = "credhub_setup/cc/testhelpers",
+    importpath = "credhub_setup/pkg/cc/testhelpers",
     visibility = ["//visibility:public"],
+    deps = ["//pkg/quarks/testhelpers:go_default_library"],
 )
