+++ deploy/containers/credhub_setup/pkg/credhub/doc.go
@@ -1,0 +1,2 @@
+// Package credhub includes the code specific to handling credhub links.
+package credhub
