+++ deploy/containers/credhub_setup/pkg/quarks/resolver.go
