+++ deploy/containers/credhub_setup/pkg/quarks/testhelpers/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["fakemount.go"],
+    importpath = "credhub_setup/pkg/quarks/testhelpers",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/quarks:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
+    ],
+)
