--- deploy/containers/credhub_setup/quarks/testhelpers/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["fakemount.go"],
-    importpath = "credhub_setup/quarks/testhelpers",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_stretchr_testify//assert:go_default_library"],
-)
